[WIP] 멀티 플레이 기능 개발중 게임 타이머 설정 게임 로직 구현

+++ YogaYo/domain/src/main/java/com/d104/domain/model/PeerUser.kt
@@ -4,6 +4,6 @@ data class PeerUser(
     val id: String,
     val nickName: String,
     val isReady: Boolean = false,
-    val totalScore: Int = 0,
+    val totalScore: Float = 0.0f,
     val roundScore: Float = 0.0f
 )
